Dispatch Update 1.0 & Hacking Update 0.5
Add police alerts which have a higher chance of being activated when close to "business" hours set in config. An alert for a suspicious activity and  thermite explosion during the higher chance time period respectively, has a 50% and a 20% chance of being activated. 
If jewellery cases are smashed without the main Vangelico's PC being hacked, a police alert will be sent out 100% of the time, but only for the first case hit in the store.
Blips were converted from strings to text commands for optimisation... ;) MrNewb